Extract Class Declaration method
This cleared up the cache code generator so that it would be easier to test
