+++ src/CacheCodeGenerator.php
@@ -11,26 +11,28 @@ class CacheCodeGenerator
      *
      * @todo clean up logic, DRY it up, maybe move
      *       some of this into Zend\Code
-     * @param  ClassReflection $r
+     * @param  ClassReflection $classReflection
      * @return string
      */
-    public function getCacheCode(ClassReflection $r)
+    public function getCacheCode(ClassReflection $classReflection)
     {
         $fileReflectionService = new FileReflectionUseStatementService();
 
-        $useStatementDto = $fileReflectionService->getUseStatementDto($r->getDeclaringFile());
+        $classDeclarationService = new ClassDeclarationService();
+
+        $useStatementDto = $fileReflectionService->getUseStatementDto($classReflection->getDeclaringFile());
 
         $useString = $useStatementDto->getUseString();
         $useNames = $useStatementDto->getUseNames();
 
-        $declaration = $this->getClassDeclaration($r, $useNames);
+        $declaration = $classDeclarationService->getClassDeclaration($classReflection, $useNames);
 
-        $classContents = $r->getContents(false);
-        $classFileDir  = dirname($r->getFileName());
+        $classContents = $classReflection->getContents(false);
+        $classFileDir  = dirname($classReflection->getFileName());
         $classContents = trim(str_replace('__DIR__', sprintf("'%s'", $classFileDir), $classContents));
 
         $return = "\nnamespace "
-            . $r->getNamespaceName()
+            . $classReflection->getNamespaceName()
             . " {\n"
             . $useString
             . $declaration . "\n"
@@ -38,68 +40,5 @@ class CacheCodeGenerator
             . "\n}\n";
 
         return $return;
-    }
-
-    /**
-     * @param ClassReflection $r
-     * @param $useNames
-     * @return string
-     */
-    private function getClassDeclaration(ClassReflection $r, $useNames)
-    {
-        $declaration = '';
-
-        if ($r->isAbstract() && !$r->isInterface()) {
-            $declaration .= 'abstract ';
-        }
-
-        if ($r->isFinal()) {
-            $declaration .= 'final ';
-        }
-
-        if ($r->isInterface()) {
-            $declaration .= 'interface ';
-        }
-
-        if (!$r->isInterface()) {
-            $declaration .= 'class ';
-        }
-
-        $declaration .= $r->getShortName();
-
-        $parentName = false;
-        if (($parent = $r->getParentClass()) && $r->getNamespaceName()) {
-            $parentName = array_key_exists($parent->getName(), $useNames)
-                ? ($useNames[$parent->getName()] ? : $parent->getShortName())
-                : ((0 === strpos($parent->getName(), $r->getNamespaceName()))
-                    ? substr($parent->getName(), strlen($r->getNamespaceName()) + 1)
-                    : '\\' . $parent->getName());
-        } else if ($parent && !$r->getNamespaceName()) {
-            $parentName = '\\' . $parent->getName();
-        }
-
-        if ($parentName) {
-            $declaration .= " extends {$parentName}";
-        }
-
-        $interfaces = array_diff($r->getInterfaceNames(), $parent ? $parent->getInterfaceNames() : array());
-        if (count($interfaces)) {
-            foreach ($interfaces as $interface) {
-                $iReflection = new ClassReflection($interface);
-                $interfaces = array_diff($interfaces, $iReflection->getInterfaceNames());
-            }
-            $declaration .= $r->isInterface() ? ' extends ' : ' implements ';
-            $declaration .= implode(', ', array_map(function ($interface) use ($useNames, $r) {
-                $iReflection = new ClassReflection($interface);
-                return (array_key_exists($iReflection->getName(), $useNames)
-                    ? ($useNames[$iReflection->getName()] ? : $iReflection->getShortName())
-                    : ((0 === strpos($iReflection->getName(), $r->getNamespaceName()))
-                        ? substr($iReflection->getName(), strlen($r->getNamespaceName()) + 1)
-                        : '\\' . $iReflection->getName()));
-            }, $interfaces));
-            return $declaration;
-        }
-
-        return $declaration;
     }
 }
